Compiled basic worker libevent implementation
