+++ src/include/wire/libevent_server.h
@@ -29,39 +29,40 @@
 
 #include "common/logger.h"
 #include "common/config.h"
-#include "wire/wire.h"
 #include "container/lock_free_queue.h"
-#include "wire/libevent_socket.h"
 #include <sys/file.h>
 #include <fstream>
+#include "wire/wire.h"
 
+#define SOCKET_BUFFER_SIZE 8192
 #define QUEUE_SIZE 100
 
-// TODO: Rename to libevent_server.h
 namespace peloton {
-
 namespace wire {
 
-struct LibeventServer {
- private:
-  // For logging purposes
-  static void LogCallback(int severity, const char* msg);
-  uint64_t port_;             // port number
-  size_t max_connections_;  // maximum number of connections
+/* Libevent Callbacks */
 
- public:
-  LibeventServer();
-  static inline LibeventSocket* GetConn(const int& connfd);
-  static inline void CreateNewConn(const int& connfd, short ev_flags,
-                                   std::shared_ptr<LibeventThread>& thread);
+/* Used by a worker thread to receive a new connection from the main thread and
+ * launch the event handler */
+void WorkerHandleNewConn(evutil_socket_t local_fd, short ev_flags, void *arg);
 
- private:
-  /* Maintain a global list of connections.
-   * Helps reuse connection objects when possible
-   */
-  static inline
-    std::vector<std::unique_ptr<LibeventSocket>>& GetGlobalSocketList();
+/* Used by a worker to execute the main event loop for a connection */
+void EventHandler(evutil_socket_t connfd, short ev_flags, void *arg);
 
+// Buffers used to batch messages at the socket
+struct Buffer {
+  size_t buf_ptr;   // buffer cursor
+  size_t buf_size;  // buffer size
+  SockBuf buf;
+
+  inline Buffer() : buf_ptr(0), buf_size(0) {}
+
+  inline void Reset() {
+    buf_ptr = 0;
+    buf_size = 0;
+  }
+
+  inline size_t GetMaxSize() { return SOCKET_BUFFER_SIZE; }
 };
 
 struct NewConnQueueItem {
@@ -90,8 +91,7 @@ class LibeventThread {
   /* The queue for new connection requests */
   LockFreeQueue<std::shared_ptr<NewConnQueueItem>> new_conn_queue;
 
-
-public:
+ public:
   inline ~LibeventThread() {}
 
   static void ProcessConnRequest(int num);
@@ -101,14 +101,88 @@ public:
   LibeventThread() : new_conn_queue(QUEUE_SIZE) {}
 };
 
-  /* Libevent Callbacks */
 
-  /* Used by a worker thread to receive a new connection from the main thread and
-   * launch the event handler */
-  void WorkerHandleNewConn(evutil_socket_t local_fd, short ev_flags, void *arg);
+/*
+ * SocketManager - Wrapper for managing socket.
+ * 	B is the STL container type used as the protocol's buffer.
+ */
+class LibeventSocket {
+ public:
+  int sock_fd;  // socket file descriptor
+  bool is_disconnected; // is the connection disconnected
+  struct event *event; // libevent handle
+  short event_flags;  // event flags mask
+  Buffer rbuf;  // Socket's read buffer
+  Buffer wbuf;  // Socket's write buffer
+  LibeventThread *thread; // reference to the libevent thread
+  std::unique_ptr<PacketManager> pkt_manager; // Stores state for this socket
 
-  /* Used by a worker to execute the main event loop for a connection */
-  void EventHandler(evutil_socket_t connfd, short ev_flags, void *arg);
+ private:
+  /* refill_read_buffer - Used to repopulate read buffer with a fresh
+  * batch of data from the socket
+  */
+  bool RefillReadBuffer();
+
+  inline void Init(short event_flags, LibeventThread *thread) {
+    is_disconnected = false;
+    this->event_flags = event_flags;
+    this->thread = thread;
+    event = event_new(thread->libevent_base_, sock_fd, event_flags, EventHandler, nullptr);
+    event_add(event, nullptr);
+  }
+
+ public:
+  inline LibeventSocket(int sock_fd, short event_flags,
+                        LibeventThread *thread) :
+      sock_fd(sock_fd), event_flags(event_flags), thread(thread) {
+    Init(event_flags, thread);
+  }
+
+
+  // Reads a packet of length "bytes" from the head of the buffer
+  bool ReadBytes(PktBuf &pkt_buf, size_t bytes);
+
+  // Writes a packet into the write buffer
+  bool BufferWriteBytes(PktBuf &pkt_buf, size_t len, uchar type);
+
+  void PrintWriteBuffer();
+
+  // Used to invoke a write into the Socket, once the write buffer is ready
+  bool FlushWriteBuffer();
+
+  void CloseSocket();
+
+  /* Resuse this object for a new connection. We could be assigned to a
+   * new thread, change thread reference.
+   */
+  void Reset(short event_flags, LibeventThread *thread) {
+    is_disconnected = false;
+    rbuf.Reset();
+    wbuf.Reset();
+    pkt_manager.reset(nullptr);
+    Init(event_flags, thread);
+  }
+};
+
+struct LibeventServer {
+ private:
+  // For logging purposes
+  static void LogCallback(int severity, const char* msg);
+  uint64_t port_;             // port number
+  size_t max_connections_;  // maximum number of connections
+
+ public:
+  LibeventServer();
+  static LibeventSocket* GetConn(const int& connfd);
+  static void CreateNewConn(const int& connfd, short ev_flags, LibeventThread *thread);
+
+ private:
+  /* Maintain a global list of connections.
+   * Helps reuse connection objects when possible
+   */
+  static std::vector<std::unique_ptr<LibeventSocket>>& GetGlobalSocketList();
+
+};
 
 }
 }
